mlflow-deploy/requirements.txt
mlflow==2.15.1
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.4.2
pyyaml==6.0.1
joblib==1.4.2
